aether-os/exports/astroapp1/requirements.txt
openai
anthropic
ollama
python-dotenv
pyyaml
